Finishing touch to the UpcomingHomeworkViewController
